+++ src/main/java/com/publicholidaysbycountry/initializer/HolidayInitializer.java
@@ -2,7 +2,9 @@ package com.publicholidaysbycountry.initializer;
 
 import com.publicholidaysbycountry.country.application.CountryService;
 import com.publicholidaysbycountry.country.domain.Country;
+import com.publicholidaysbycountry.holiday.application.HolidayApiService;
 import com.publicholidaysbycountry.holiday.application.HolidayService;
+import com.publicholidaysbycountry.holiday.domain.Holiday;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -19,13 +21,16 @@ import org.springframework.stereotype.Component;
 public class HolidayInitializer {
 
     private final CountryService countryService;
+    private final HolidayApiService holidayApiService;
     private final HolidayService holidayService;
 
     @EventListener(ApplicationReadyEvent.class)
     public void initializeHolidays() {
         List<Country> countries = countryService.saveCountries();
         log.info("국가 데이터 일괄 적재 완료: 국가 {}개 저장", countries.size());
-        int savedHolidayCount = holidayService.saveHolidays(countries, LocalDate.now().getYear());
+
+        List<Holiday> holidays = holidayApiService.getHolidaysFromApiByCountryAndYear(countries, LocalDate.now().getYear());
+        int savedHolidayCount = holidayService.saveHolidays(holidays);
         log.info("공휴일 데이터 일괄 적재 완료: 공휴일 {}개 저장", savedHolidayCount);
     }
 }
